Consolidate React.cache tests into basic.test.ts and update server implementation
